feat(comment): 优化评论列表接口并添加评论树功能
- 修改 CommentMapper 接口，将 page 方法返回类型改为 List<CommentRes>- 在 CommentDao 中添加 count 方法，用于获取评论总数
- 在 CommentRes 中添加 parentId、rootParentId、replyNickName 和 children 字段
- 实现评论树功能，将子评论挂载到父评论的 children 列表中- 优化 CommentService 接口，使用 PageResult 封装分页数据
- 在 ArticleServiceImpl 中添加更新 Redis 缓存的逻辑
- 调整 BaseContext 和 CacheUtil 类，支持新的缓存操作

+++ blog-back/blog-web/src/main/java/com/yyjy/web/domain/vo/request/CommentPageReq.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class CommentPageReq {
 	private long articleId;
-	private int current = 1;
-	private int size = 5;
+	private long current = 1L;
+	private long size = 5L;
 }
